[ZEPPELIN-3089] Delete all notes for current user
"Empty Trash" is not working when multi-user create a notebook and move it to trash.
Because all notebooks are in the same trash folder even though the user doesn't have permission to the notebook, and on clicking "Empty Trash" button next to Trash folder, Zeppelin tries to delete all notebooks in the Trash folder, which includes no permission notebook, but cannot delete those because of no permission.

This has slightly different implementation than that is there is https://github.com/apache/zeppelin/pull/2695

[Bug Fix]

* [ ] - Add unit test case

* [ZEPPELIN-3089](https://issues.apache.org/jira/browse/ZEPPELIN-3089)

* Login as user 'A' and create a notebook 'A'
* Move notebook 'A' to Trash
* Login as user 'B' and create a notebook 'B'
* Move notebook 'B' to Trash
* Click 'Empty Trash' button next to 'Trash' folder

![zeppelin-3249-python](https://user-images.githubusercontent.com/674497/37451922-229533ea-2859-11e8-9a76-085c5cb8b9d1.gif)

* Does the licenses files need update? N/A
* Is there breaking changes for older versions? N/A
* Does this needs documentation? N/A

Author: Prabhjyot Singh <prabhjyotsingh@gmail.com>

Closes #2872 from prabhjyotsingh/ZEPPELIN-3089 and squashes the following commits:

971cbf672 [Prabhjyot Singh] ZEPPELIN-3089: Delete all notes for current user

Change-Id: I88f0028e5a7859d7c767c3b3ea58dc0cad1aaa43