[ZEPPELIN-3399] Tests failing with no changes - ShellScriptLauncherTest and SparkInterpreterLaunchTest
### What is this PR for?
When I run all of the tests zeppelin-zengine in IntelliJ, the tests in ShellScriptLauncherTest and SparkInterpreterLaunchTest fail, but if I run these tests by themselves, they pass.
There is a problem with the tests in ShellScripterLauncherTest and SparkInterpreterLaunchTest, which is that if some of the ConfVars are set (such as ZEPPELIN_HOME) in other tests, then these tests will fail, because they expect that these variables are not set.

To fix this, I have added a setUp() method that clears all of the ConfVars to both ShellScripterLauncherTest and SparkInterpreterLauncherTest. This way, none of the ConfVars will be set when these tests are run.

### What type of PR is it?
Bug Fix for tests ShellScripterLauncherTest and SparkInterpreterLauncherTest.

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3399

### How should this be tested?
Can verify tests still work after pull request by running them, and can verify that they failed before by running all tests in IntelliJ in the default order that IntelliJ runs them.

Author: Reed Oei <oei.reed@gmail.com>

Closes #2918 from ReedOei/ShellScriptLauncherTest-testLauncher-fix and squashes the following commits:

6f57ab4 [Reed Oei] Fixed tests in SparkInterpreterLauncherTest.
bcbbabd [Reed Oei] Fixed test: ShellScripterLauncherTest.testLauncher.